requirements.txt
Django==5.2.12
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
django-cors-headers==4.9.0
psycopg[binary]==3.3.3
PyJWT[crypto]==2.11.0
boto3==1.42.66
uWSGI==2.0.31
